Remove allocate deploy step.
